+++ fit-grid-layout/fit-grid-layout-loader.js
@@ -6,7 +6,7 @@
 import { HAControlLoader } from "../ha-control-loader.js?v=0.6.7";
 
 // Cache-busting version parameter for script loading
-const VERSION = "1.0.0";
+const VERSION = "1.0.1";
 
 // Name of this loader module script
 const SCRIPT_NAME = "fit-grid-layout-loader.js";
@@ -17,5 +17,5 @@ const loader = new HAControlLoader(SCRIPT_NAME, VERSION);
 // Dynamically load stylesheets and scripts needed by the view layout
 loader.loadModules(
   ["fit-grid-layout.css"],
-  ["fit-grid-layout.js"]
+  ["fit-grid-layout.js", "fit-grid-layout-editor.js"]
 );